% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breve_GetBaseClusters.R
\name{breve_GetBaseClusters}
\alias{breve_GetBaseClusters}
\title{Apply multiple clustering methods on an -omics dataset to predict base clusters.}
\usage{
breve_GetBaseClusters(
  selected_data,
  params,
  clustering_methods = c("NEMO", "PINS", "SNF", "kMeans")
)
}
\arguments{
\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{clustering_methods}{a vector of valid clustering methods.}
}
\value{
a data.frame associating samples to their predicted clusters.
Its rows are samples, its columns are clustering methods, and predicted populations are reported in the table.
}
\description{
Apply multiple clustering methods on an -omics dataset to predict base clusters.
}
