% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breve_GetCharacteristicFeatures.R
\name{breve_GetCharacteristicFeatures}
\alias{breve_GetCharacteristicFeatures}
\title{Get differentially expressed features by calling the function \code{FindMarkers} of the edgeR package.
Only genes with abs(logFC) > 2 and FDR < 0.001 are returned.}
\usage{
breve_GetCharacteristicFeatures(
  cluster,
  selected_data,
  params,
  logFC_threshold = 2,
  pvalue_threshold = 0.001
)
}
\arguments{
\item{cluster}{a named lists, with five names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{logFC_threshold}{a numeric.}

\item{pvalue_threshold}{a numeric.}
}
\value{
a named vector associating over-expressed features to their log2 fold-changes.
}
\description{
Get differentially expressed features by calling the function \code{FindMarkers} of the edgeR package.
Only genes with abs(logFC) > 2 and FDR < 0.001 are returned.
}
