% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breve_GetSelectedFeatures.R
\name{breve_GetSelectedFeatures}
\alias{breve_GetSelectedFeatures}
\title{Get the variable features in an -omics dataset.}
\usage{
breve_GetSelectedFeatures(dataset, params, n_features = 1000)
}
\arguments{
\item{dataset}{an -omics dataset.
Its rows are features and its columns are samples.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{n_features}{the number of highly variable features to sample.}
}
\value{
a vector of genes.
}
\description{
The variable features have a variance superior to 0.
}
