% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_characterized_clusters.R
\name{draw_characteristic_features}
\alias{draw_characteristic_features}
\title{Get an upset-plot representing the characteristic features predicted in each meta-cluster.}
\usage{
draw_characteristic_features(characterized_clusters)
}
\arguments{
\item{characterized_clusters}{a list where every element is a characterized pool of samples.
The elements are named lists, with six names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label}, \code{features} and \code{robustness}.}
}
\value{
a plot.
}
\description{
Get an upset-plot representing the characteristic features predicted in each meta-cluster.
}
