% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{draw_robust_clusters}
\alias{draw_robust_clusters}
\title{Get composite U-MAP plots representing the meta-clusters predicted with the fEVE framework.}
\usage{
draw_robust_clusters(robust_clusters, selected_data)
}
\arguments{
\item{robust_clusters}{a list where every element is a pool of samples.
The elements are named lists, with five names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}
}
\value{
a plot.
}
\description{
Get composite U-MAP plots representing the meta-clusters predicted with the fEVE framework.
}
