% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_data.R
\name{draw_selected_data}
\alias{draw_selected_data}
\title{Get a U-MAP plot representing the pool of samples used in the clustering recursion.}
\usage{
draw_selected_data(population, SeuratObject_init, records)
}
\arguments{
\item{population}{a character. It corresponds to the cell population that scEVE will attempt to cluster.}

\item{SeuratObject_init}{a SeuratObject generated from dataset_init, on which
the function RunUMAP() of Seurat has been applied already.}

\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}
}
\value{
a plot.
}
\description{
Get a U-MAP plot representing the pool of samples used in the clustering recursion.
}
