% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feve.R
\name{feve}
\alias{feve}
\title{Conduct a clustering analysis with the fEVE framework, starting from the initial population C.}
\usage{
feve(dataset_init, params, figures = TRUE, sheets = TRUE)
}
\arguments{
\item{dataset_init}{a dataset, without selected features.
Its rows are features and its columns are samples.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering recursion.}

\item{sheets}{a boolean that indicates if the results of the clustering recursion should be saved in Excel sheets.}
}
\value{
a named list, with two elements: \code{records} and \code{preds}.
\code{records} is a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.
\code{preds} is a named factor associating samples to their predicted clusters.
}
\description{
Conduct a clustering analysis with the fEVE framework, starting from the initial population C.
}
