% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feve_HardClustering.R
\name{feve_HardClustering}
\alias{feve_HardClustering}
\title{Get a table reporting the composition of the characterized clusters.}
\usage{
feve_HardClustering(characterized_clusters, selected_data, params)
}
\arguments{
\item{characterized_clusters}{a list where every element is a characterized pool of samples.
The elements are named lists, with seven names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label}, \code{features}, \code{robustness} and \code{specific_features}.}

\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\value{
a data.frame associating samples to their predicted clusters.
Its rows are samples, its columns are characterized clusters, and cluster memberships are reported in the table.
}
\description{
By calling this function, the samples are hard clustered,
i.e. their membership to their assigend cluster is 1, and their membership to the other clusters is 0.
}
