% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feve.R
\name{feve_recursion}
\alias{feve_recursion}
\title{Attempt to cluster a specific population using the fEVE algorithm.}
\usage{
feve_recursion(
  population,
  dataset_init,
  SeuratObject_init,
  records,
  params,
  figures,
  sheets
)
}
\arguments{
\item{population}{a character. It corresponds to the population that scEVE will attempt to cluster.}

\item{dataset_init}{a dataset, without selected features.
Its rows are features and its columns are samples.}

\item{SeuratObject_init}{a SeuratObject generated from dataset_init, on which
the function RunUMAP() of Seurat has been applied already.}

\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering recursion.}

\item{sheets}{a boolean that indicates if the results of the clustering recursion should be saved in Excel sheets.}
}
\value{
a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.
}
\description{
Attempt to cluster a specific population using the fEVE algorithm.
}
