% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{filter_conflicting_samples}
\alias{filter_conflicting_samples}
\title{Filter out conflicting samples belonging to multiple robust clusters.}
\usage{
filter_conflicting_samples(robust_clusters)
}
\arguments{
\item{robust_clusters}{a list where every element is a pool of samples.
The elements are named lists, with five names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label} and \code{robustness}.}
}
\value{
a list where every element is a pool of samples.
The elements are named lists, with five names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label} and \code{robustness}.
}
\description{
If a sample exists in multiple robust clusters, it is exclusively associated
to its most robust cluster. Robust clusters without samples are also removed.
}
