% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_plot.R
\name{get_barplot_compositions}
\alias{get_barplot_compositions}
\title{Get a barplot associating each leaf cluster to its sample composition.}
\usage{
get_barplot_compositions(tree_data, ground_truth, records_samples)
}
\arguments{
\item{tree_data}{a data structure suitable for \code{ggtree::ggtree()}.}

\item{ground_truth}{a named factor associating samples to their ground truth label.}

\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and and its columns are populations. The cell values range from 0 to 1.}
}
\value{
a plot.
}
\description{
Get a barplot associating each leaf cluster to its sample composition.
}
