% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_plot.R
\name{get_barplot_features}
\alias{get_barplot_features}
\title{Get a barplot associating each leaf cluster to its features.}
\usage{
get_barplot_features(tree_data, records_features)
}
\arguments{
\item{tree_data}{a data structure suitable for \code{ggtree::ggtree()}.}

\item{records_features}{a data.frame associating predicted populations to their characteristic features.
Its rows are genes, its columns are predicted populations,
and the strength of the characterization (e.g. log2-transformed pvalues) are reported in the table.}
}
\value{
a plot.
}
\description{
Get a barplot associating each leaf cluster to its features.
}
