% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_contributions.R
\name{get_barplot_robustness}
\alias{get_barplot_robustness}
\title{Get a barplot associating each robust cluster to its robustness value.}
\usage{
get_barplot_robustness(records_meta)
}
\arguments{
\item{records_meta}{a data.frame associating predicted populations to generic information, including:
their \code{size}, their \code{robustness}, their \code{parent} and their \code{clustering_status}.}
}
\value{
a plot.
}
\description{
Get a barplot associating each robust cluster to its robustness value.
}
