% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_summary.R
\name{get_barplot_sizes}
\alias{get_barplot_sizes}
\title{Get a barplot associating each leaf cluster to its size.}
\usage{
get_barplot_sizes(tree_data, records_samples)
}
\arguments{
\item{tree_data}{a data structure suitable for \code{ggtree::ggtree()}.}

\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and its columns are population. The cell values range from 0 to 1.}
}
\value{
a plot.
}
\description{
Get a barplot associating each leaf cluster to its size.
}
