% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_clusters.R
\name{get_base_clusters}
\alias{get_base_clusters}
\title{Get base clusters predicted with multiple clustering methods.}
\usage{
get_base_clusters(population, selected_data, params, figures)
}
\arguments{
\item{population}{a character. It corresponds to the population that feature will attempt to cluster.}

\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering recursion.}
}
\value{
a data.frame associating samples to their predicted clusters.
Its rows are samples, its columns are clustering methods, and predicted populations are reported in the table.
}
\description{
Get base clusters predicted with multiple clustering methods.
}
