% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark.R
\name{get_benchmark}
\alias{get_benchmark}
\title{Using computational as well as intrinsic and extrinsic clustering metrics, measure
the performance of an instance of the fEVE framework on a dataset.}
\usage{
get_benchmark(data, params, method_label = "fEVE")
}
\arguments{
\item{data}{a named list with two elements: \code{dataset} and \code{ground_truth}.
\code{dataset} is a dataset, without selected features. Its rows are features and its columns are samples.
\code{ground_truth} is a named factor associating samples to their cluster annotations.}

\item{params}{a list of parameters (cf. \code{feve::get_default_parameters()}).}

\item{method_label}{a character.}
}
\value{
a data.frame with eight columns: \code{method}, \code{time (s)},
\code{peak_memory_usage (Mb)}, \code{ARI}, \code{NMI}, \code{nPurity}, \code{SI} and \code{n_samples}.
}
\description{
Using computational as well as intrinsic and extrinsic clustering metrics, measure
the performance of an instance of the fEVE framework on a dataset.
}
