% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_plot.R
\name{get_cluster_compositions}
\alias{get_cluster_compositions}
\title{Get a data.frame associating each cluster to its class composition,
according to some ground truth.}
\usage{
get_cluster_compositions(ground_truth, records_samples)
}
\arguments{
\item{ground_truth}{a named factor associating samples to their ground truth label.}

\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and and its columns are populations. The cell values range from 0 to 1.}
}
\value{
a data.frame with three columns: \code{cluster}, \code{sample_class} and \code{n}.
}
\description{
Get a data.frame associating each cluster to its class composition,
according to some ground truth.
}
