% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_plot.R
\name{get_cluster_features}
\alias{get_cluster_features}
\title{Get a data.frame associating each cluster to its features.}
\usage{
get_cluster_features(records_features)
}
\arguments{
\item{records_features}{a data.frame associating predicted populations to their characteristic features.
Its rows are genes, its columns are predicted populations,
and the strength of the characterization (e.g. log2-transformed pvalues) are reported in the table.}
}
\value{
a data.frame with three columns: \code{cluster}, \code{n} and \code{names}.
}
\description{
Get a data.frame associating each cluster to its features.
}
