% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_summary.R
\name{get_cluster_sizes}
\alias{get_cluster_sizes}
\title{Get a data.frame associating each cluster to its size.}
\usage{
get_cluster_sizes(records_samples)
}
\arguments{
\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and its columns are population. The cell values range from 0 to 1.}
}
\value{
a data.frame with two columns: \code{cluster} and \code{n}.
}
\description{
Get a data.frame associating each cluster to its size.
}
