% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clustering_metrics.R
\name{get_clustering_metrics}
\alias{get_clustering_metrics}
\title{Using both intrinsic and extrinsic clustering metrics, measure a clustering performance.}
\usage{
get_clustering_metrics(data, preds)
}
\arguments{
\item{data}{a named list with two elements: \code{dataset} and \code{ground_truth}.
\code{dataset} is a dataset, without selected features. Its rows are features and its columns are samples.
\code{ground_truth} is a named factor associating samples to their cluster annotations.}

\item{preds}{a named vector associating cells to their predicted clusters.}
}
\value{
a named vector with four names: \code{ARI}, \code{NMI}, \code{nPurity} and \code{SI}.
}
\description{
Extrinsic metrics compare cluster predictions to the cell annotations of the dataset.
Intrinsic metrics compare the gene expression of cells in and out of their clusters.
The \code{ARI} and the \code{NMI} are extrinsic metrics.
The \code{nPurity} and the \code{SI} are intrinsic metrics.
For every metric, higher is better and the maximum value is 1.
}
