% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clustering_metrics.R
\name{get_clustering_metrics_trifecta}
\alias{get_clustering_metrics_trifecta}
\title{Using both intrinsic and extrinsic clustering metrics, measure the clustering performances
of a fEVE clustering (with and without its leftover clusters), and the ground truth.}
\usage{
get_clustering_metrics_trifecta(data, preds, method_label)
}
\arguments{
\item{data}{a named list with two elements: \code{dataset} and \code{ground_truth}.
\code{dataset} is a dataset, without selected features. Its rows are features and its columns are samples.
\code{ground_truth} is a named factor associating samples to their cluster annotations.}

\item{preds}{a named vector associating cells to their predicted clusters.}

\item{method_label}{a character.}
}
\value{
a data.frame with six columns: \code{method}, \code{ARI}, \code{NMI}, \code{nPurity}, \code{SI} and \code{n_samples}.
}
\description{
Using both intrinsic and extrinsic clustering metrics, measure the clustering performances
of a fEVE clustering (with and without its leftover clusters), and the ground truth.
}
