% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark.R
\name{get_data_bluster}
\alias{get_data_bluster}
\title{Get a matrix usable for the functions of bluster.}
\usage{
get_data_bluster(dataset)
}
\arguments{
\item{dataset}{a dataset, without selected features.
Its rows are features and its columns are samples.}
}
\value{
an object associating samples and selected features to their reduced dimensions.
}
\description{
Get a matrix usable for the functions of bluster.
}
