% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_GetBaseClusters.R
\name{get_data_input}
\alias{get_data_input}
\title{Get the data input relevant to a clustering method.}
\usage{
get_data_input(method, data)
}
\arguments{
\item{method}{a clustering method used by scEVE.
Valid methods include: densityCut, monocle3, Seurat and SHARP.}

\item{data}{a named list, with three names: \code{dataset}, \code{SeuratObject} and \code{logtpm}.
They correspond to the scRNA-seq expression matrix of a specific cell population, its SeuratObject and its log2-transformed TPM values.}
}
\value{
one of \code{data$dataset}, \code{data$SeuratObject} or \code{data$logtpm}.
}
\description{
Get the data input relevant to a clustering method.
}
