% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feve_GetCharacteristicFeatures.R
\name{get_differential_feature}
\alias{get_differential_feature}
\title{Get the log2 fold change, as well as the FDR, of a feature.}
\usage{
get_differential_feature(feature, filtered_dataset, groups)
}
\arguments{
\item{feature}{a character.}

\item{filtered_dataset}{an -omics dataset, with variant features.
Its rows are features and its columns are samples.}

\item{groups}{an ordered vector of integers.}
}
\value{
a data.frame with three columns: \code{feature}, \code{logFC} and \code{pval}.
}
\description{
Get the log2 fold change, as well as the FDR, of a feature.
}
