% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metheve_GetCharacteristicFeatures.R
\name{get_differential_methylation}
\alias{get_differential_methylation}
\title{Get the log2 fold change, as well as the FDR, of a methylation site.}
\usage{
get_differential_methylation(site, filtered_dataset, groups)
}
\arguments{
\item{site}{a character.}

\item{filtered_dataset}{an -omics dataset, with variant features.
Its rows are features and its columns are samples.}

\item{groups}{an ordered vector of integers.}
}
\value{
a data.frame with three columns: \code{site}, \code{logFC} and \code{pval}.
}
\description{
Get the log2 fold change, as well as the FDR, of a methylation site.
}
