% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recorded_recursion.R
\name{get_drawings_paths}
\alias{get_drawings_paths}
\title{Get the paths leading to each informative figure drawn during the clustering recursion.}
\usage{
get_drawings_paths(population, params)
}
\arguments{
\item{population}{a character. It corresponds to the population that fEVE will attempt to cluster.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\value{
a vector of paths.
}
\description{
Get the paths leading to each informative figure drawn during the clustering recursion.
}
