% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breve_GetBaseClusters.R,
%   R/sceve_GetBaseClusters.R
\name{get_formatted_predictions}
\alias{get_formatted_predictions}
\title{Format the cluster predictions of a method to obtain a standard output.}
\usage{
get_formatted_predictions(samples, predictions)

get_formatted_predictions(samples, predictions)
}
\arguments{
\item{samples}{a vector of sample names.}

\item{predictions}{a vector of cluster predictions.}
}
\value{
a named factor that associates each sample to their cluster prediction.

a named factor that associates each sample to their cluster prediction.
}
\description{
Format the cluster predictions of a method to obtain a standard output.

Format the cluster predictions of a method to obtain a standard output.
}
