% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_contributions.R
\name{get_heatmap_contributions}
\alias{get_heatmap_contributions}
\title{Get a heatmap plot representing which method contributed to the prediction of each cluster.}
\usage{
get_heatmap_contributions(records_methods)
}
\arguments{
\item{records_methods}{a data.frame associating methods (rows) to the robust clusters
they predicted (columns). Cell values are binary, and 1 indicates that a method contributed
in the detection of a robust cluster.}
}
\value{
a plot
}
\description{
Get a heatmap plot representing which method contributed to the prediction of each cluster.
}
