% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_benchmark.R
\name{get_intrinsic_clustering_metrics}
\alias{get_intrinsic_clustering_metrics}
\title{Using intrinsic metrics, measure a clustering performance.}
\usage{
get_intrinsic_clustering_metrics(dataset, preds)
}
\arguments{
\item{dataset}{a dataset, without selected features.
Its rows are features and its columns are samples.}

\item{preds}{a named factor associating cells to their predicted clusters.}
}
\value{
a named vector of with two names: \code{nPurity} and \code{SI}.
}
\description{
Using intrinsic metrics, measure a clustering performance.
}
