% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_leaf_clusters}
\alias{get_leaf_clusters}
\title{Get a named vector associating each sample to its most informative cluster label.}
\usage{
get_leaf_clusters(records_samples)
}
\arguments{
\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and its columns are population. The cell values range from 0 to 1.}
}
\value{
a named vector associating samples to their most informative cluster labels.
}
\description{
Get a named vector associating each sample to its most informative cluster label.
}
