% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_leaf_clusters_at_resolution}
\alias{get_leaf_clusters_at_resolution}
\title{Get a named vector associating each sample to its most informative cluster label, at a given maximum resolution.}
\usage{
get_leaf_clusters_at_resolution(resolution, records_samples)
}
\arguments{
\item{resolution}{an integer.}

\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and its columns are population. The cell values range from 0 to 1.}
}
\value{
a named vector associating samples to their most informative cluster labels, at a given maximum resolution.
}
\description{
Get a named vector associating each sample to its most informative cluster label, at a given maximum resolution.
}
