% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_maximum_resolution}
\alias{get_maximum_resolution}
\title{Get the maximum clustering resolution attained in a fEVE clustering analysis.}
\usage{
get_maximum_resolution(records_samples)
}
\arguments{
\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and its columns are populations. The cell values range from 0 to 1.}
}
\value{
an integer.
}
\description{
Get the maximum clustering resolution attained in a fEVE clustering analysis.
}
