% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feve.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters to conduct a fEVE clustering analysis.}
\usage{
get_parameters(params_label)
}
\arguments{
\item{params_label}{a character.

These parameters are:
\code{random_state} the integer seed used to have deterministic results.
\code{minimum_samples} the minimum number of samples expected in a population before attempting to sub-cluster it.
\code{figures_path} & \code{sheets_path} the default paths where figures and result sheets are stored, respectively.
\code{selected_features_strategy} a function called to select a limited pool of features for a clustering recursion.
\code{base_clusters_strategy} a function called to predict base clusters with multiple clustering methods.
\code{characteristic_features_strategy} a function called to predict the characteristic features of every meta-cluster.
\code{characterized_clusters_strategy} a function called to identify characterized clusters.
\code{cluster_memberships_strategy} a function called to assign ambiguous samples to their clusters (i.e. hard or soft-clustering).

Detailed information regarding the strategy parameters are available in the vignette of the package.}
}
\value{
a list of parameters.
}
\description{
Get parameters to conduct a fEVE clustering analysis.
}
