% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feve.R
\name{get_pending_population}
\alias{get_pending_population}
\title{Get a population for which no fEVE clustering recursion has been attempted.}
\usage{
get_pending_population(records)
}
\arguments{
\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}
}
\value{
a character.
}
\description{
Get a population for which no fEVE clustering recursion has been attempted.
}
