% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_report.R
\name{get_placeholder_ground_truth}
\alias{get_placeholder_ground_truth}
\title{Get a placeholder ground truth corresponding to shuffled fEVE labels.}
\usage{
get_placeholder_ground_truth(records, random_state = 1)
}
\arguments{
\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}

\item{random_state}{a numeric.}
}
\value{
a named factor associating samples to their ground truth label.
}
\description{
Get a placeholder ground truth corresponding to shuffled fEVE labels.
}
