% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_contributions.R
\name{get_plot_contributions}
\alias{get_plot_contributions}
\title{Get a composite plot representing the contribution of each clustering method
on the fEVE prediction.}
\usage{
get_plot_contributions(records, heights = c(1, 1))
}
\arguments{
\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}

\item{heights}{a vector of numeric.}
}
\value{
a composite plot
}
\description{
Get a composite plot representing the contribution of each clustering method
on the fEVE prediction.
}
