% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_resolutions.R
\name{get_plot_resolutions}
\alias{get_plot_resolutions}
\title{Get lineplots summarizing the cluster quality of a fEVE prediction across multiple resolutions.}
\usage{
get_plot_resolutions(data, records_samples, method_label = "fEVE")
}
\arguments{
\item{data}{a named list with two elements: \code{dataset} and \code{ground_truth}.
\code{dataset} is a dataset, without selected features. Its rows are features and its columns are samples.
\code{ground_truth} is a named factor associating samples to their cluster annotations.}

\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and its columns are population. The cell values range from 0 to 1.}

\item{method_label}{a character.}
}
\value{
a faceted plot.
}
\description{
Get lineplots summarizing the cluster quality of a fEVE prediction across multiple resolutions.
}
