% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analysis_report.R
\name{get_plot_wrapper}
\alias{get_plot_wrapper}
\title{Wrapper function to generate and save a plot.}
\usage{
get_plot_wrapper(plot_label, data, records)
}
\arguments{
\item{plot_label}{a character.}

\item{data}{a named list with two elements: \code{dataset} and \code{ground_truth}.
\code{dataset} is a dataset, without selected features. Its rows are features and its columns are samples.
\code{ground_truth} is a named factor associating samples to their cluster annotations.}

\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}
}
\value{
a named list with two names: \code{plot} and \code{dimensions}.
}
\description{
Wrapper function to generate and save a plot.
}
