% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recorded_recursion.R
\name{get_recorded_recursion}
\alias{get_recorded_recursion}
\title{Report multiple information regarding the characterized clusters predicted during a clustering recursion.}
\usage{
get_recorded_recursion(
  population,
  characterized_clusters,
  selected_data,
  records,
  params
)
}
\arguments{
\item{population}{a character. It corresponds to the population that fEVE will attempt to cluster.}

\item{characterized_clusters}{a list where every element is a characterized pool of samples.
The elements are named lists, with seven names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label}, \code{features}, \code{robustness} and \code{specific_features}.}

\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{records}{a named list, with three data.frames: \code{samples}, \code{features} and \code{meta}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\value{
a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.
}
\description{
Report multiple information regarding the characterized clusters predicted during a clustering recursion.
}
