% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_resolutions.R
\name{get_resolution_data}
\alias{get_resolution_data}
\title{Get the clustering metrics at each resolution to be plotted.}
\usage{
get_resolution_data(data, records_samples, method_label)
}
\arguments{
\item{data}{a named list with two elements: \code{dataset} and \code{ground_truth}.
\code{dataset} is a dataset, without selected features. Its rows are features and its columns are samples.
\code{ground_truth} is a named factor associating samples to their cluster annotations.}

\item{records_samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and its columns are population. The cell values range from 0 to 1.}

\item{method_label}{a character.}
}
\value{
a data.frame with five columns: \code{method}, \code{n_samples}, \code{resolution}, \code{metric} and \code{value}.
}
\description{
Get the clustering metrics at each resolution to be plotted.
}
