% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{get_robust_clusters}
\alias{get_robust_clusters}
\title{Extract robust clusters and a leftover cluster from a set of base clusters predicted
with multiple clustering methods.}
\usage{
get_robust_clusters(
  population,
  base_clusters,
  selected_data,
  records,
  params,
  figures
)
}
\arguments{
\item{population}{a character. It corresponds to the population that fEVE will attempt to cluster.}

\item{base_clusters}{a data.frame associating samples to their predicted clusters.
Its rows are samples, its columns are clustering methods, and predicted populations are reported in the table.}

\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{figures}{a boolean that indicates if figures should be drawn to explain the clustering recursion.}
}
\value{
a list where every element is a pool of samples.
The elements are named lists, with five names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label} and \code{robustness}.
}
\description{
An empty list is returned if no robust cluster was predicted.
}
