% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_robust_clusters.R
\name{get_robustness_threshold}
\alias{get_robustness_threshold}
\title{Get the robustness threshold to identify robust clusters.}
\usage{
get_robustness_threshold(population, base_clusters, records)
}
\arguments{
\item{population}{a character. It corresponds to the population that fEVE will attempt to cluster.}

\item{base_clusters}{a data.frame associating samples to their predicted clusters.
Its rows are samples, its columns are clustering methods, and predicted populations are reported in the table.}

\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}
}
\value{
a numeric.
}
\description{
Get the robustness threshold to identify robust clusters.
}
