% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_samples_of_population}
\alias{get_samples_of_population}
\title{Get every sample belonging in a predicted population.}
\usage{
get_samples_of_population(population, records.samples)
}
\arguments{
\item{population}{a character. It corresponds to a population predicted by fEVE (e.g. C.2).}

\item{records.samples}{a data.frame associating samples to their predicted populations.
Its rows are samples and and its columns are population. The cell values range from 0 to 1.}
}
\value{
a vector of samples.
}
\description{
Get every sample belonging in a predicted population.
}
