% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_data.R
\name{get_selected_dataset}
\alias{get_selected_dataset}
\title{Get a dataset with selected features.}
\usage{
get_selected_dataset(dataset, params)
}
\arguments{
\item{dataset}{a dataset, without selected features.
Its rows are features and its columns are samples.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\value{
a dataset, with selected features.
}
\description{
Get a dataset with selected features.
}
