% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_data.R
\name{get_selected_SeuratObject}
\alias{get_selected_SeuratObject}
\title{Get a SeuratObject from a dataset with selected features.}
\usage{
get_selected_SeuratObject(selected_dataset, params)
}
\arguments{
\item{selected_dataset}{a dataset, with selected features.
Its rows are features and its columns are samples.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\value{
a SeuratObject, on which the function ScaleData() of Seurat has been applied already.
}
\description{
Get a SeuratObject from a dataset with selected features.
}
