% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breve_GetBaseClusters.R
\name{get_transformed_dataset}
\alias{get_transformed_dataset}
\title{Get a log-transformed and normalized dataset.}
\usage{
get_transformed_dataset(filtered_dataset)
}
\arguments{
\item{filtered_dataset}{an -omics dataset, with variant features.
Its rows are features and its columns are samples.}
}
\value{
a log-transformed and normalized -omics dataset, with variant features.
Its rows are features and its columns are samples.
}
\description{
Get a log-transformed and normalized dataset.
}
