% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_plot.R
\name{get_tree_data}
\alias{get_tree_data}
\title{Get a data structure representing the hierarchy of clusters predicted.
The robustness of each cluster is also included.
The resulting data structure is suitable for \code{ggtree::ggtree()}.}
\usage{
get_tree_data(records_meta)
}
\arguments{
\item{records_meta}{associating predicted populations to generic information, including:
their \code{size}, their \code{robustness}, their \code{parent} and their \code{clustering_status}.}
}
\value{
a data structure suitable for \code{ggtree::ggtree()}.
}
\description{
Get a data structure representing the hierarchy of clusters predicted.
The robustness of each cluster is also included.
The resulting data structure is suitable for \code{ggtree::ggtree()}.
}
