% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_summary.R
\name{get_tree_plot}
\alias{get_tree_plot}
\title{Get a plot representing the hierarchy of clusters predicted.}
\usage{
get_tree_plot(tree_data)
}
\arguments{
\item{tree_data}{a data structure suitable for \code{ggtree::ggtree()}.}
}
\value{
a plot.
}
\description{
Get a plot representing the hierarchy of clusters predicted.
}
