% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feve.R
\name{initialize_records}
\alias{initialize_records}
\title{Get a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}
\usage{
initialize_records(dataset_init)
}
\arguments{
\item{dataset_init}{a dataset, without selected features.
Its rows are features and its columns are samples.}
}
\value{
a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.
}
\description{
\itemize{
\item \code{samples} associates samples to their predicted populations.
Its rows are samples, its columns are predicted populations, and cluster memberships are reported in the table.
\item \code{features} associates predicted populations to their characteristic features.
Its rows are features, its columns are predicted populations, and characterization powers are reported in the table.
\item \code{meta} associates predicted populations to generic information, including:
their \code{size}, their \code{robustness}, their \code{parent} and their \code{clustering_status}.
\item \code{methods} associates predicted populations to the clustering methods leveraged to predict them.
Its rows are clustering methods, its columns are predicted populations, and binary values are reported in the table.
}
}
