% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recorded_recursion.R
\name{merge_drawings}
\alias{merge_drawings}
\title{Merge the informative figures drawn during the clustering recursion in a single .pdf file.}
\usage{
merge_drawings(population, params)
}
\arguments{
\item{population}{a character. It corresponds to the population that fEVE will attempt to cluster.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\description{
Merge the informative figures drawn during the clustering recursion in a single .pdf file.
}
