% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recorded_recursion.R
\name{report_metadata}
\alias{report_metadata}
\title{Report the metadata regarding characterized clusters in the corresponding records sheet.}
\usage{
report_metadata(population, characterized_clusters, records)
}
\arguments{
\item{population}{a character. It corresponds to the population that fEVE will attempt to cluster.}

\item{characterized_clusters}{a list where every element is a characterized pool of samples.
The elements are named lists, with seven names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label}, \code{features}, \code{robustness} and \code{specific_features}.}

\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}
}
\value{
a data.frame associating predicted populations to generic information, including:
their \code{size}, their \code{robustness}, their \code{parent} and their \code{clustering_status}.
}
\description{
Report the metadata regarding characterized clusters in the corresponding records sheet.
}
