% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_GetBaseClusters.R
\name{sceve_GetBaseClusters}
\alias{sceve_GetBaseClusters}
\title{Apply multiple clustering methods on scRNA-seq data to predict base clusters.}
\usage{
sceve_GetBaseClusters(
  selected_data,
  params,
  clustering_methods = c("densityCut", "monocle3", "Seurat", "SHARP")
)
}
\arguments{
\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{clustering_methods}{a vector of valid clustering methods.}
}
\value{
a data.frame associating cells to their predicted clusters.
Its rows are cells, its columns are clustering methods, and predicted populations are reported in the table.
}
\description{
Apply multiple clustering methods on scRNA-seq data to predict base clusters.
}
