% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_GetCharacteristicFeatures.R
\name{sceve_GetCharacteristicFeatures}
\alias{sceve_GetCharacteristicFeatures}
\title{Get marker genes by calling the function \code{FindMarkers} of the Seurat package.
Only genes with log2FC > 4 and p-values (corrected with Bonferonni) < 0.001 are returned.}
\usage{
sceve_GetCharacteristicFeatures(
  cluster,
  selected_data,
  params,
  logFC_threshold = 4,
  pvalue_threshold = 0.001
)
}
\arguments{
\item{cluster}{a named lists, with five names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label} and \code{robustness}.}

\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{logFC_threshold}{a numeric.}

\item{pvalue_threshold}{a numeric.}
}
\value{
a named vector associating marker genes to their log2 fold-changes.
}
\description{
Get marker genes by calling the function \code{FindMarkers} of the Seurat package.
Only genes with log2FC > 4 and p-values (corrected with Bonferonni) < 0.001 are returned.
}
