% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_GetCharacterizedClusters.R
\name{sceve_GetCharacterizedClusters}
\alias{sceve_GetCharacterizedClusters}
\title{Filter out uncharacterized clusters, i.e. clusters with too little marker genes.}
\usage{
sceve_GetCharacterizedClusters(
  clusters,
  selected_data,
  params,
  features_threshold = 10
)
}
\arguments{
\item{clusters}{a list where every element is a pool of samples.
The elements are named lists, with six names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label}, \code{features} and \code{robustness}.}

\item{selected_data}{a named list, with two names: \code{dataset} and \code{SeuratObject}.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{features_threshold}{the minimal number of marker features expected in a characterized cluster.}
}
\value{
a list where every element is a pool of samples.
The elements are named lists, with six names:
\code{base_clusters}, \code{samples}, \code{clustering_methods}, \code{label}, \code{features} and \code{robustness}.
}
\description{
Filter out uncharacterized clusters, i.e. clusters with too little marker genes.
}
