% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_LoadData.R
\name{sceve_GetDatasets}
\alias{sceve_GetDatasets}
\title{Get information regarding the datasets available on TMExplorer for scEVE.}
\usage{
sceve_GetDatasets()
}
\value{
a data.frame associating each dataset to its sequencing protocol,
its number of cells, clusters and genes, as well as its year of publication,
its accession number and its associated doi.
}
\description{
Get information regarding the datasets available on TMExplorer for scEVE.
}
