% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_GetSelectedFeatures.R
\name{sceve_GetSelectedFeatures}
\alias{sceve_GetSelectedFeatures}
\title{Get the n most variable genes in a scRNA-seq dataset.}
\usage{
sceve_GetSelectedFeatures(dataset, params, n_features = 1000)
}
\arguments{
\item{dataset}{a scRNA-seq dataset of raw count dataset, without selected genes.
Its rows are genes and its columns are cells.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{n_features}{the number of highly variable genes to sample.}
}
\value{
a vector of genes.
}
\description{
The variable genes are identified by calling the function FindVariableFeatures() of Seurat.
}
