% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_LoadData.R
\name{sceve_LoadData}
\alias{sceve_LoadData}
\title{Load a scRNA-seq dataset of raw count expression.}
\usage{
sceve_LoadData(dataset_label)
}
\arguments{
\item{dataset_label}{one of \code{Li_HumCRC_b}, \code{Darmanis_HumGBM}, \code{JerbyArnon_HumMLM},
\code{Gillen_HumEPN}, \code{VanGalen_HumAML}, \code{Lambrechts_HumNSCLC} or \code{Peng_HumPDAC}.
(cf. \code{sceve::get_datasets()})}
}
\value{
a named list with two elements: \code{dataset} and \code{ground_truth}.
\code{dataset} is a scRNA-seq dataset of raw count expression, without selected genes.
Its rows are genes and its columns are cells.
\code{ground_truth} is a named factor associating cells to their cluster annotations.
}
\description{
Load a scRNA-seq dataset of raw count expression.
}
