% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_GetBaseClusters.R
\name{use_densityCut}
\alias{use_densityCut}
\title{Predict clusters with the densityCut method.}
\usage{
use_densityCut(logtpm_dataset, params)
}
\arguments{
\item{logtpm_dataset}{a scRNA-seq dataset of log-transformed transcripts per millions (tpm),
with selected genes.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\value{
a named factor that associates each cell to their cluster prediction.
}
\description{
Predict clusters with the densityCut method.
}
