% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceve_GetBaseClusters.R
\name{use_monocle3}
\alias{use_monocle3}
\title{Predict clusters with the monocle3 method.}
\usage{
use_monocle3(selected_SeuratObject, params)
}
\arguments{
\item{selected_SeuratObject}{a SeuratObject, on which the function ScaleData()
of Seurat has been applied already.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\value{
a named factor that associates each cell to their cluster prediction.
}
\description{
Predict clusters with the monocle3 method.
}
