% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breve_GetBaseClusters.R
\name{use_PINS}
\alias{use_PINS}
\title{Predict clusters with the PINS method.}
\usage{
use_PINS(filtered_dataset, params)
}
\arguments{
\item{filtered_dataset}{an -omics dataset, with variant features.
Its rows are features and its columns are samples.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}
}
\value{
a named factor that associates each patient to their cluster prediction.
}
\description{
Predict clusters with the PINS method.
}
