% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_analysis_report.R
\name{write_analysis_report}
\alias{write_analysis_report}
\title{Save summary figures of an analysis in a .pdf file.}
\usage{
write_analysis_report(
  dataset_init,
  records,
  ground_truth = NULL,
  path = "./report.pdf"
)
}
\arguments{
\item{dataset_init}{a dataset, without selected features.
Its rows are features and its columns are samples.}

\item{records}{a named list, with four data.frames: \code{samples}, \code{features}, \code{meta} and \code{methods}.}

\item{ground_truth}{a named factor associating samples to their ground truth label.}

\item{path}{a character.}
}
\description{
Save summary figures of an analysis in a .pdf file.
}
