% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feve_GetSelectedFeatures.R
\name{feve_GetSelectedFeatures}
\alias{feve_GetSelectedFeatures}
\title{Get the variable features in an -omics dataset.}
\usage{
feve_GetSelectedFeatures(dataset, params, features_percent = 0.25)
}
\arguments{
\item{dataset}{an -omics dataset.
Its rows are features and its columns are samples.}

\item{params}{a list of parameters (cf. \code{feve::get_parameters()}).}

\item{features_percent}{the proportion of highly variable features to select
(according to the initial pool of features).}
}
\value{
a vector of features.
}
\description{
The variable features have a variance superior to 0.
}
